% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exclusionlists}
\alias{exclusionlists}
\title{Mitochondrial exclusionlist}
\format{
A list with three entries:
#' \itemize{
  \item \emph{three}: Regions of the mitochondrial genome that are within 1 nt of a 3-mer homopolymer (e.g. AAA)
  \item \emph{mutaseq}: Mutations in the mitochondrial genome that were reoccuring across patients (present in more than one individual in the MutaSeq dataset)
  \item \emph{masked}: Regions of the mitochondrial genome that are soft-masked in the UCSC or Ensembl annotations
  \item \emph{rnaEDIT}: Regions of the mitochondrial genome that are thought to be subject to RNA-editing according to the REDIportal V2.0
}
}
\usage{
exclusionlists
}
\description{
List of variants that are likely not true somatic mutations and should thus be excluded
}
\keyword{datasets}
