% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getAlleleCount}
\alias{getAlleleCount}
\title{mutationCalls counts accessor}
\usage{
getAlleleCount(mutcall, type = c("mutant", "nonmutant"))
}
\arguments{
\item{mutcall}{object of class \code{\link{mutationCalls}}.}

\item{type}{character that is either `mutant` or `nonmutant`
depending on which allele count the user wants to access}
}
\value{
Returns matrix of either mutant or non-mutant allele counts
}
\description{
Extracts the counts of allele for either the mutant or all the
non-mutant alleles
}
\examples{
load(system.file("extdata/LudwigFig7.Rda",package = "mitoClone2"))
mutantAllele_count <- getAlleleCount(LudwigFig7,type='mutant')
}
