% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{getVarsCandidate}
\alias{getVarsCandidate}
\title{mutationCalls cluster accessor}
\usage{
getVarsCandidate(mutcall)
}
\arguments{
\item{mutcall}{object of class \code{\link{mutationCalls}}.}
}
\value{
Returns a character vector including all the variants to be
used for clustering
}
\description{
Extracts all the putative variants that we want to use for
clustering
}
\examples{
load(system.file("extdata/LudwigFig7.Rda",package =
"mitoClone2"))
mutations_to_cluster <- getVarsCandidate(LudwigFig7)
}
