% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{mut2GR}
\alias{mut2GR}
\title{Convert mutation string to GRanges}
\usage{
mut2GR(mut)
}
\arguments{
\item{mut}{The mutation to convert to a GRanges in the format of "position reference>alternate".}
}
\value{
Returns a GRanges object containg the site of the variant along with reference/alternate allele data in the metacolumns
}
\description{
Convert mutation string to GRanges
}
\examples{
mutation_as_granges <- mut2GR('1434 G>A')
mutation_as_granges_no_space <- mut2GR('1434G>A')
}
