% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plotTree}
\alias{plotTree}
\title{Plot clonal tree}
\usage{
plotTree(mutcalls, file = "mytree.ps")
}
\arguments{
\item{mutcalls}{object of class \code{\link{mutationCalls}}.}

\item{file}{file name of postscript output file}
}
\value{
Returns TRUE only used for generating a PostScript tree image of the putative mutation tree
}
\description{
Uses graphviz to plot the clonal tree
}
\examples{
load(system.file("extdata/LudwigFig7.Rda",package = "mitoClone2"))
LudwigFig7.plot <- plotTree(LudwigFig7)
}
