% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCalling.R
\name{pullcountsVars}
\alias{pullcountsVars}
\title{Pull variant counts}
\usage{
pullcountsVars(BaseCounts, vars, cells = NULL)
}
\arguments{
\item{BaseCounts}{A list of base call matrices (one matrix per cell) as produced by \code{\link{baseCountsFromBamList}}}

\item{vars}{Character vector of variants to pull, in format 5643G>T}

\item{cells}{Character vector for cells to select, or NULL if all cells from the input are to be used}
}
\value{
A list with two entries, M (count table on the variant allele) and N (count table on the reference allele)
}
\description{
Pull variant counts
}
\examples{
LudwigFig7.Counts <- readRDS(url("http://steinmetzlab.embl.de/mutaseq/fig7_nucleotide_counts_per_position.RDS"))
known.variants <- c("9000 T>C","1234 G>A","1337 G>A")
counts.known.vars <- pullcountsVars(LudwigFig7.Counts, vars=known.variants)
}
