% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vireo.R
\name{vireo}
\alias{vireo}
\title{Initialize Vireo model}
\usage{
vireo(
  n.cell,
  n.var,
  n.donor,
  n.gt = 3,
  learn.gt = TRUE,
  fix.beta.sum = FALSE,
  beta.mu.init = NULL,
  beta.sum.init = NULL,
  id.prob.init = NULL,
  gt.prob.init = NULL
)
}
\arguments{
\item{n.cell}{Number of cells}

\item{n.var}{Number of variants}

\item{n.donor}{Number of donors}

\item{n.gt}{Number of genotype states (default 3: 0,1,2)}

\item{learn.gt}{Whether to learn genotype probabilities}

\item{fix.beta.sum}{Whether to fix beta sum parameters}

\item{beta.mu.init}{Initial beta mu values}

\item{beta.sum.init}{Initial beta sum values}

\item{id.prob.init}{Initial ID probabilities}

\item{gt.prob.init}{Initial genotype probabilities}
}
\description{
Initialize Vireo model
}
